pysyslog/pysyslog/requirements.txt
torch
transformers
pika
python-dotenv
stomp-py
sentence-transformers